improvement: changes the interface for AnswerClassifier (#4)
* improvement: change the interface for AnswerClassifier

* fixes: format
